Moviment

+++ castleofillusion/TileMap.cpp
@@ -140,7 +140,7 @@ void TileMap::prepareArrays(const glm::vec2 &minCoords, ShaderProgram &program) 
 // Method collisionMoveDown also corrects Y coordinate if the box is
 // already intersecting a tile below.
 
-bool TileMap::collisionMoveLeft(const glm::ivec2 &pos, const glm::ivec2 &size) const
+bool TileMap::collisionMoveLeft(const glm::vec2 &pos, const glm::vec2 &size) const
 {
 	int x, y0, y1;
 	
@@ -154,7 +154,7 @@ bool TileMap::collisionMoveLeft(const glm::ivec2 &pos, const glm::ivec2 &size) c
 	return false;
 }
 
-bool TileMap::collisionMoveRight(const glm::ivec2 &pos, const glm::ivec2 &size) const {
+bool TileMap::collisionMoveRight(const glm::vec2 &pos, const glm::vec2 &size) const {
 	int x, y0, y1;
 	
 	x = (pos.x + size.x - 1) / tileSize;
@@ -167,7 +167,7 @@ bool TileMap::collisionMoveRight(const glm::ivec2 &pos, const glm::ivec2 &size) 
 	return false;
 }
 
-bool TileMap::collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, int *posY) const {
+bool TileMap::collisionMoveDown(const glm::vec2 &pos, const glm::vec2 &size, float *posY) const {
 	int x0, x1, y;
 	
 	x0 = pos.x / tileSize;
